--- a/main/src/core/MonoDevelop.Core/MonoDevelop.Projects.MSBuild/MSBuildEvaluationContext.cs
+++ b/main/src/core/MonoDevelop.Core/MonoDevelop.Projects.MSBuild/MSBuildEvaluationContext.cs
@@ -44,12 +44,12 @@ namespace MonoDevelop.Projects.MSBuild
 	sealed class MSBuildEvaluationContext: IExpressionContext
 	{
 		Dictionary<string,string> properties = new Dictionary<string, string> (StringComparer.OrdinalIgnoreCase);
-		static Dictionary<string, string> envVars = new Dictionary<string, string> ();
-		readonly HashSet<string> propertiesWithTransforms = new HashSet<string> (StringComparer.OrdinalIgnoreCase);
-		readonly List<string> propertiesWithTransformsSorted = new List<string> ();
+		readonly IDictionary envVars;
+		readonly HashSet<string> propertiesWithTransforms;
+		readonly List<string> propertiesWithTransformsSorted;
 		List<ImportSearchPathExtensionNode> searchPaths;
 
-		public Dictionary<string, bool> ExistsEvaluationCache { get; } = new Dictionary<string, bool> ();
+		public Dictionary<string, bool> ExistsEvaluationCache { get; }
 
 		bool allResolved;
 		MSBuildProject project;
@@ -65,16 +65,22 @@ namespace MonoDevelop.Projects.MSBuild
 
 		public MSBuildEvaluationContext ()
 		{
+			ExistsEvaluationCache = new Dictionary<string, bool> ();
+			propertiesWithTransforms = new HashSet<string> (StringComparer.OrdinalIgnoreCase);
+			propertiesWithTransformsSorted = new List<string> ();
+			envVars = Environment.GetEnvironmentVariables ();
 		}
 
 		public MSBuildEvaluationContext (MSBuildEvaluationContext parentContext)
 		{
 			this.parentContext = parentContext;
 			this.project = parentContext.project;
+			this.Log = parentContext.Log;
+
+			this.ExistsEvaluationCache = parentContext.ExistsEvaluationCache;
 			this.propertiesWithTransforms = parentContext.propertiesWithTransforms;
 			this.propertiesWithTransformsSorted = parentContext.propertiesWithTransformsSorted;
-			this.ExistsEvaluationCache = parentContext.ExistsEvaluationCache;
-			this.Log = parentContext.Log;
+			this.envVars = parentContext.envVars;
 		}
 
 		internal void InitEvaluation (MSBuildProject project)
@@ -277,12 +283,7 @@ namespace MonoDevelop.Projects.MSBuild
 			if (parentContext != null)
 				return parentContext.GetPropertyValue (name);
 
-			lock (envVars) {
-				if (!envVars.TryGetValue (name, out val))
-					envVars[name] = val = Environment.GetEnvironmentVariable (name);
-
-				return val;
-			}
+			return (string)envVars [name];
 		}
 
 		public string GetMetadataValue (string name)