--- a/main/tests/test-projects/msbuild-project-test/test.csproj
+++ b/main/tests/test-projects/msbuild-project-test/test.csproj
@@ -68,5 +68,7 @@
   </ItemGroup>
   <Target Name="Conditioned" Condition="false"></Target>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
+  <!-- Imported twice for verification -->
+  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="extra.csproj" />
 </Project>