TODO: Figure out whether GetImportFiles/ImportFile can be optimized in any way.

Not too big gains, but quick ones, for around 30MB on solution load. Containing code still allocates around 140MB (was 170).