@Therzok 2018-03-27
`ProjectLoadSaveTests` sets an EnvironmentVariable at the beginning of the test. If any other test hits evaluation before that happens.

Added do not merge for the test until I profile it again to see perf stats.