@Therzok 2018-03-27
The other thing that could be done is invalidate the env var cache via the tests, but that feels hacky.