@Therzok 2018-03-28
This is good to go. Just verified, performance is good.