@Therzok 2018-03-29
Old timings:
```
100.00%   Evaluate  •  34,967 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(Object, Boolean)
  99.94%   EvaluateProject  •  34,945 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.EvaluateProject(ProjectInfo, MSBuildEvaluationContext)
    98.49%   EvaluateObjects  •  34,439 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.EvaluateObjects(ProjectInfo, MSBuildEvaluationContext, IEnumerable[MSBuildObject], Boolean)
      45.88%   Evaluate  •  16,044 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildImport, Boolean)
        ►22.73%   ImportFile  •  7,947 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.ImportFile(ProjectInfo, MSBuildEvaluationContext, MSBuildImport, String, String)
        ►18.44%   GetImportFiles  •  6,447 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.GetImportFiles(ProjectInfo, MSBuildEvaluationContext, MSBuildImport, String, String, String&, Boolean&)
         1.79%   IndexOf  •  626 ms  •  System.String.IndexOf(String, StringComparison)
        ►1.16%   EvaluateProject  •  405 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.EvaluateProject(ProjectInfo, MSBuildEvaluationContext, Boolean)
        ►1.13%   Append  •  396 ms  •  MonoDevelop.Projects.ConditionedPropertyCollection.Append(ConditionedPropertyCollection)
         0.13%   Concat  •  45 ms  •  System.String.Concat(String, String, String)
        ►0.11%   MSBuildEvaluationContext..ctor  •  37 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext..ctor(MSBuildEvaluationContext)
         0.10%   Add  •  34 ms  •  System.Collections.Generic.List`1.Add(T)
         0.06%   GetEnumerator  •  21 ms  •  System.Collections.Generic.List`1.GetEnumerator
         0.05%   get_Property  •  16 ms  •  MonoDevelop.Projects.Extensions.ImportSearchPathExtensionNode.get_Property
        ►0.04%   GetImportedProjects  •  13 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.GetImportedProjects(ProjectInfo, MSBuildImport)
         0.03%   GetProjectImportSearchPaths  •  11 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext.GetProjectImportSearchPaths
      ►35.20%   Evaluate  •  12,308 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildPropertyGroup)
      ►6.68%   Evaluate  •  2,334 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildImportGroup, Boolean)
      ►5.44%   Evaluate  •  1,901 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildTarget)
      ►3.34%   Evaluate  •  1,169 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildItemGroup)
      ►1.31%   Evaluate  •  456 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildChoose, Boolean)
       0.35%   MoveNext  •  122 ms  •  System.Linq.Enumerable+<OfTypeIterator>d__93`1.MoveNext
    ►0.97%   InitEvaluation  •  339 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext.InitEvaluation(MSBuildProject)
    ►0.34%   EvaluateWithItems  •  121 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext.EvaluateWithItems(String, List[MSBuildItemEvaluated])
     0.08%   Exists  •  28 ms  •  System.IO.File.Exists(String)
     0.05%   TryGetValue  •  18 ms  •  System.Collections.Generic.Dictionary`2.TryGetValue(TKey, TValue&)
  ►0.03%   SetPropertyValue  •  11 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext.SetPropertyValue(String, String)
   0.03%   get_Current  •  11 ms  •  System.Collections.Generic.Dictionary+Enumerator`2.get_Current
```

New timings:
```
100.00%   Evaluate  •  19,672 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(Object, Boolean)
  99.47%   EvaluateProject  •  19,568 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.EvaluateProject(ProjectInfo, MSBuildEvaluationContext)
    98.40%   EvaluateObjects  •  19,357 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.EvaluateObjects(ProjectInfo, MSBuildEvaluationContext, IEnumerable[MSBuildObject], Boolean)
      ►42.77%   Evaluate  •  8,414 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildPropertyGroup)
      36.56%   Evaluate  •  7,193 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildImport, Boolean)
        ►21.15%   GetImportFiles  •  4,160 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.GetImportFiles(ProjectInfo, MSBuildEvaluationContext, MSBuildImport, String, String, String&, Boolean&)
        ►10.91%   ImportFile  •  2,146 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.ImportFile(ProjectInfo, MSBuildEvaluationContext, MSBuildImport, String, String)
        ►1.37%   Append  •  270 ms  •  MonoDevelop.Projects.ConditionedPropertyCollection.Append(ConditionedPropertyCollection)
         1.05%   IndexOf  •  206 ms  •  System.String.IndexOf(String, Int32, StringComparison)
        ►0.91%   EvaluateProject  •  180 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.EvaluateProject(ProjectInfo, MSBuildEvaluationContext, Boolean)
         0.39%   Add  •  78 ms  •  System.Collections.Generic.List`1.Add(T)
         0.12%   IndexOf  •  24 ms  •  System.String.IndexOf(String, StringComparison)
        ►0.08%   GetImportedProjects  •  17 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.GetImportedProjects(ProjectInfo, MSBuildImport)
         0.08%   MSBuildEvaluationContext..ctor  •  15 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext..ctor(MSBuildEvaluationContext)
         0.04%   set_IsImported  •  7 ms  •  MonoDevelop.Projects.MSBuild.MSBuildItemEvaluated.set_IsImported(Boolean)
      ►6.23%   Evaluate  •  1,225 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildItemGroup)
      ►5.11%   Evaluate  •  1,006 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildTarget)
      ►4.00%   Evaluate  •  787 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildImportGroup, Boolean)
      ►2.06%   Evaluate  •  405 ms  •  MonoDevelop.Projects.MSBuild.DefaultMSBuildEngine.Evaluate(ProjectInfo, MSBuildEvaluationContext, MSBuildChoose, Boolean)
       0.86%   MoveNext  •  169 ms  •  System.Linq.Enumerable+<OfTypeIterator>d__93`1.MoveNext
    ►0.91%   InitEvaluation  •  178 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext.InitEvaluation(MSBuildProject)
     0.16%   Exists  •  32 ms  •  System.IO.File.Exists(String)
  ►0.41%   MSBuildEvaluationContext..ctor  •  81 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext..ctor
  ►0.06%   SetPropertyValue  •  11 ms  •  MonoDevelop.Projects.MSBuild.MSBuildEvaluationContext.SetPropertyValue(String, String)
```