@Therzok 2018-03-29
Current bottleneck is a new lock contention on DefaultMSBuildEngine.conditionCache. To investigate perf gains for locking per condition file or having an immutable dictionary.