@Therzok 2018-03-29
Load before this PR
![oldload](https://user-images.githubusercontent.com/109974/38096167-e51d2ca2-337a-11e8-906b-955da42b81b6.gif)

Load after this PR
![newload](https://user-images.githubusercontent.com/109974/38096420-7ace02ee-337b-11e8-8a61-4ca9b649953b.gif)